Following current style & structure proposals
